refactor: finish route migration
