+++ package.json
@@ -4,11 +4,11 @@
   "version": "0.1.0",
   "type": "module",
   "scripts": {
-    "dev": "svelte-kit dev",
-    "build": "pnpm build:icons && svelte-kit build",
+    "dev": "vite dev",
+    "build": "pnpm build:icons && vite build",
     "build:icons": "node src/sharp.js",
-    "package": "svelte-kit package",
-    "preview": "svelte-kit preview",
+    "package": "vite package",
+    "preview": "vite preview",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
     "lint": "prettier --check --plugin-search-dir=. . && eslint .",
