+++ src/routes/(book)/+layout.svelte
