+++ src/routes/(book)/[slug]/+page@(book).svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
-	import type { Page } from "./[slug]@book";
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
+	import type { Page } from "../[slug]@book";
 
 	export let page: Page;
 </script>
