+++ src/routes/(book)/notebook/+page@(book).svelte
