+++ src/routes/(book)/words/+page.server.ts
@@ -1,3 +1,5 @@
+throw new Error("@migration task: Update +page.server.js (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292699)");
+
 import { resolve } from "node:path";
 import { readdirSync, readFileSync } from "node:fs";
