+++ src/routes/(book)/words/+page@(book).svelte
@@ -1,4 +1,6 @@
 <script type="ts">
+	throw new Error("@migration task: Add data prop (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292707)");
+
 	import type { Post } from "./index@book";
 
 	const formatted = (date: Date): `${string}.${string}` =>
