+++ src/routes/(book)/words/[slug]/+page.server.ts
@@ -1,10 +1,12 @@
+throw new Error("@migration task: Update +page.server.js (https://github.com/sveltejs/kit/discussions/5774#discussioncomment-3292699)");
+
 import { existsSync, readFileSync } from "node:fs";
 import { resolve } from "node:path";
 
 import grayMatter from "gray-matter";
 import { marked } from "marked";
-import type { RequestHandler } from ".svelte-kit/types/src/routes/words/__types/[slug]@book";
-import type { Post } from "./index@book";
+import type { RequestHandler } from "..svelte-kit/types/src/routes/words/__types/[slug]@book";
+import type { Post } from "../index@book";
 
 export const get: RequestHandler = async ({ params }) => {
 	const { slug } = params;
