+++ src/routes/+layout.svelte
@@ -1,4 +1,4 @@
-<script>
+<script lang="ts">
 	import "@fontsource/nunito/400.css";
 	import "@fontsource/nunito/600.css";
 </script>
