--- svelte.config.js
@@ -1,12 +1,0 @@
-import adapter from "@sveltejs/adapter-auto";
-import preprocess from "svelte-preprocess";
-
-/** @type {import('@sveltejs/kit').Config} */
-const config = {
-	preprocess: preprocess(),
-	kit: {
-		adapter: adapter()
-	}
-}
-
-export default config;
