+++ vite.config.js
@@ -1,0 +1,8 @@
+import { sveltekit } from "@sveltejs/kit/vite";
+
+/** @type {import('vite').UserConfig} */
+const config = {
+	plugins: [sveltekit()],
+};
+
+export default config;
